+++ tests/tests_integration/test_api/test_simulators/test_integrations.py
@@ -27,8 +27,10 @@ class TestSimulatorIntegrations:
 
         assert len(all_integrations) > 0
         assert filtered_integrations[0].external_id == seed_resource_names["simulator_integration_external_id"]
-        # check time difference
+        assert filtered_integrations[0].data_set_id == seed_resource_names["simulator_test_data_set_id"]
         assert filtered_integrations[0].active is True
+        assert filtered_integrations[0].created_time is not None
+        assert filtered_integrations[0].last_updated_time is not None
 
         assert len(active_integrations) > 0
         assert len(filtered_integrations) > 0
@@ -36,27 +38,17 @@ class TestSimulatorIntegrations:
     def test_delete_integrations(self, cognite_client: CogniteClient, seed_resource_names) -> None:
         simulator_integration["heartbeat"] = int(time.time() * 1000)
         simulator_integration["externalId"] = random_string(50)
+        simulator_integration["dataSetId"] = seed_resource_names["simulator_test_data_set_id"]
 
-        cognite_client.post(
-            f"/api/v1/projects/{cognite_client.config.project}/simulators/integrations",
+        cognite_client.simulators._post(
+            "/simulators/integrations",
             json={"items": [simulator_integration]},
         )
 
-        all_integrations = cognite_client.simulators.integrations.list()
-        assert len(all_integrations) > 0
-        assert any(
-            map(
-                lambda x: x.external_id == simulator_integration["externalId"],
-                all_integrations,
-            )
-        )
+        all_integrations = cognite_client.simulators.integrations.list(limit=None)
+        assert all_integrations.get(external_id=simulator_integration["externalId"]) is not None
 
         cognite_client.simulators.integrations.delete(external_id=simulator_integration["externalId"])
 
-        all_integrations = cognite_client.simulators.integrations.list()
-        assert not any(
-            map(
-                lambda x: x.external_id == simulator_integration["externalId"],
-                all_integrations,
-            )
-        )
+        all_integrations = cognite_client.simulators.integrations.list(limit=None)
+        assert all_integrations.get(external_id=simulator_integration["externalId"]) is None