+++ tests/tests_integration/test_api/test_simulators/test_simulators.py
@@ -7,5 +7,7 @@ from cognite.client._cognite_client import CogniteClient
 class TestSimulators:
     def test_list_simulators(self, cognite_client: CogniteClient) -> None:
         simulators = cognite_client.simulators.list(limit=5)
+        for simulator in simulators:
+            assert simulator.id is not None
 
         assert len(simulators) > 0